leverage new DValue semantics
Each newly created DValue has a "refcount" of 1, i.e. it expects to
be called only once for back propagation.  This means it can immediately
backprop the gradient further.  When the same DValue is used more than
once, a buffer (reference) needs to be acquired.  The code acquiring
the buffer is also responsible for completing (releasing) it.

The upshot is that no variables need to be created for intermediate
results - these are used only once anyway.  Only for variables that
were created in the original code do buffers need to be created and
completed.